fixed level1-xml sounds and compiler warnings

+++ Breakout/SoundCollection.cpp
@@ -24,7 +24,7 @@ void SoundCollection::Clear()
 
 int SoundCollection::Size()
 {
-	return m_Sounds.size();
+	return (int)m_Sounds.size();
 }
 
 bool SoundCollection::LoadSound(string path)
